Add style.css and fix typos/punctuation
The original version of the website was plain HTML and, therefore, had
very basic look. By adding external .css file and linking it to all
pages, it looks basic no more!
Also fixed some missing punctuation in recipe pages.

IMPORTANT note: the new look of the page is not something I'd submit to
'World Championships in Web Design' competition - its main purpose is
to play around and get a basic feeling of what css is and what can be
done with it.