Change the homepage look
Previous version of the main page was just a header and a plain,
unordered list. With new knowledge about the box model under my belt,
I gave it a new look by adding some classes to existing elements,
and provided a new logo for the site. I've also added some structure
to the project by organizing all files in separate directories.

Important note: as before, this is not anywhere close to razor-sharp,
eye-candy design. The purpose of those changes is to practice the new
skills and play around with different possibilities that arise.